requirements.txt
pandas==2.0.0
spacy[cuda-autodetect]==3.5.1
openpyxl==3.1.2
jieba==0.42.1
pymongo==4.3.3
jupyterlab==3.6.3
beautifulsoup4==4.12.2
tqdm==4.65.0
pyahocorasick==2.0.0